refactor: improve auth-pages code
