+++ src/app/(auth-pages)/reset-password/page.tsx
@@ -1,5 +1,5 @@
-import { redirect } from "next/navigation"
-;
+import { redirect } from "next/navigation";
+
 import { TypographyParagraph } from "@/components/typography/paragraph";
 import { TypographyH1 } from "@/components/typography/headings";
 import { getUser } from "@/lib/auth/user";
