+++ src/components/nav-bar/auth-buttons.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 
 import { Button } from "@/components/ui/button";
 import { getUser } from "@/lib/auth/user";
-import { signOut } from "@/lib/auth/sign-out";
+import { signOut } from "@/server/sign-out";
 
 export default async function AuthButtons() {
   const user = await getUser();
